# Optimize RLE+ decoding by 20%