```
decode                  time:   [15.350 us 15.388 us 15.440 us]                    
                        change: [-19.388% -19.130% -18.910%] (p = 0.00 < 0.05)
                        Performance has improved.

decode_encode           time:   [30.458 us 30.542 us 30.659 us]                           
                        change: [-13.335% -13.014% -12.640%] (p = 0.00 < 0.05)
                        Performance has improved.
```