@Kubuxu 2022-04-02
I spent few hours learning about profiling in Rust, this is the result.