@raulk 2022-04-03
Mind elaborating on the latter comment? Are you seeing much variability, and did you run the profiling on Wasm bytecode?