@raulk 2022-04-03
Also, how invasive did you find the profiler to be?