@Kubuxu 2022-04-03
Regarding the fuel comment, in the fuel world you are incentivised to optimise for fuel not for speed.

In this case I profiled out of WASM, believing that better/faster x86 will result in better llvm-or and thus faster WASM.

Profiler only runs when you pass the —profile-time flag to criterion. Thinking about it more maybe I should add feature flag for it as people might be missing gperftools shared lib.