@Stebalien 2022-04-04
@Kubuxu when you optimized, what compiler flags did you pass? I ask because we usually use Oz/Os in actors. We _might_ need to add a few `#[inline(always)]` comments here given that we're adding a few new function calls.