@Kubuxu 2022-04-04
I used the default release profile that is in ref-fvm. I guess it might behave a bit different with Oz, I can retest.